Add files via upload
It is an on demand traffic light system. Normally allowing cars to pass, get ready, stops and so on, on 5 Seconds intervals. While an on demand mode for pedestrians can be activated to allow the pedestrians to pass the road safely by giving them the priority to pass instead of the cars.